refactor: stroage 분리
